tests/requirements.txt
appdirs==1.4.4
attrs==21.2.0
black==21.6b0
click==8.0.1
coverage==5.5
iniconfig==1.1.1
mypy==0.910
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
PyYAML==5.4.1
regex==2021.7.6
toml==0.10.2
types-PyYAML==5.4.3
typing-extensions==3.10.0.0
